pyproject.toml
[tool.black]
line-length = 110
target-version = ["py39"]

[tool.mypy]
python_version = "3.9"
strict = true

[tool.poetry]
name = "copyrighthook"
version = "0.0.1"
description = "Pre commit hook for updating copyright headers"
readme = "README.md"
keywords = ["pre-commit"]
license = "MIT"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Typing :: Typed",
]
authors = ["nanoufo <79045429+nanoufo@users.noreply.github.com>"]
repository = "https://github.com/nanoufo/copyright-hook"

[tool.poetry.scripts]
update-copyright = "copyrighthook.copyright:main"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
PyYAML = "^6.0.1"

[tool.poetry.dev-dependencies]
prospector = { version = "1.10.3", extras = ["with_bandit", "with_mypy", "with_pyroma"] }
pytest = "8.0.2"
types-PyYAML = "6.0.12.12"

[tool.poetry.group.dev.dependencies]
prospector-profile-duplicated = "^1.1.0"
types-pyyaml = "^6.0.12.12"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
